V1.2 + Tutorial Overlap
NOTE: Next commit will be moved to tutorial branch!

In this version I polished the sound setting, allowing you to change the master volume of the game.

I then worked on making the play button functional, taking you to a level select page.

Currently there is only one level - tutorial and I began making spritesheets for the character and level. (Not finished)