V1.4 Character Movement + Timer
Fixed the jump buffer and coyote jump one the character. Jumping animation for the character is still bugged. Not sure why need to seach online forum. Created a local timer with a functional record and reset ability. Need to find out how to implement across levels and speak with data file